projects/titanic/requirements.txt
marimo>=0.9.0
polars-lts-cpu>=0.20.0
altair>=5.2.0
pyarrow>=12.0.0
pandas>=1.5.0
